+++ packages/frontend/src/lib/RuleDetail.svelte
@@ -1,46 +1,70 @@
 <script lang="ts">
-  import { Mode, type Content, type JSONContent, type TextContent } from 'svelte-jsoneditor';
-  import ResponseEditor from './ResponseEditor.svelte';
-  import { ruleStore } from './stores/rules';
-  import type { Rule } from './types';
-  
-  export let selectedRule: Rule | null = null;
-  let showResponseEditor = false;
-  let editingConditionIndex = -1;
-  let responseContent: Content = { text: '' } as TextContent;
-  let mode: Mode = Mode.text;
+  import {
+    Mode,
+    type Content,
+    type JSONContent,
+    type TextContent,
+  } from "svelte-jsoneditor";
+  import ResponseEditor from "@/lib/ResponseEditor.svelte";
+  import { ruleStore } from "@/lib/stores/rules";
+  import type { Rule, RuleCondition, RuleProxyMode } from "@/lib/types";
+  import Switch from "@/lib/components/Switch.svelte";
+  import Select from "@/lib/components/Select.svelte";
+  import { LOCAL_PREFIX, PROXY_MODE } from "./context";
+  import { notifyMessage } from "@/api";
+  import { listenPageVisibility } from "./utils";
+
+  let isFirst = true;
+
+  let showResponseEditor = $state(false);
+  let editingConditionIndex = $state(-1);
+  let responseContent = $state<TextContent | JSONContent>({ text: "" });
+  let mode = $state<Mode>(Mode.text);
+
+  let selectedRule = $state<Rule | null>(null);
+
+  let conditions = $state<RuleCondition[]>([]);
+
+  ruleStore.subscribe((state) => {
+    selectedRule = state.selectedRule;
+    conditions = state.selectedRule?.config.conditions || [];
+  });
 
   function openResponseEditor(index: number) {
     editingConditionIndex = index;
     if (mode === Mode.tree) {
       try {
         responseContent = {
-          json: selectedRule?.config.conditions[index].response ? JSON.parse(selectedRule.config.conditions[index].response) : {}
+          json: selectedRule?.config.conditions[index].response
+            ? JSON.parse(selectedRule.config.conditions[index].response)
+            : {},
         } as JSONContent;
       } catch (e) {
         responseContent = { json: {} } as JSONContent;
       }
     } else {
       responseContent = {
-        text: selectedRule?.config.conditions[index].response || ''
+        text: selectedRule?.config.conditions[index].response || "",
       } as TextContent;
     }
     showResponseEditor = true;
   }
 
-  $: if (showResponseEditor === false && editingConditionIndex >= 0 && selectedRule) {
-    const config = { ...selectedRule.config };
-    // @ts-ignore
-    config.conditions[editingConditionIndex].response = mode === Mode.text ? responseContent.text : JSON.stringify(responseContent.json);
-    console.log('[info: 27]:', editingConditionIndex, responseContent, config, mode);
-    ruleStore.updateRuleConfig(selectedRule.id, config);
-    editingConditionIndex = -1;
-  }
-
   function addCondition() {
     if (!selectedRule) return;
     const config = { ...selectedRule.config };
-    config.conditions = [...config.conditions, { key: '', value: '', response: '' }];
+    config.conditions = [
+      ...config.conditions,
+      {
+        key: "",
+        value: "",
+        response: "",
+        enabled: true,
+        remark: "",
+        ruleId: selectedRule.id,
+        proxyMode: PROXY_MODE.NETWORK,
+      },
+    ];
     ruleStore.updateRuleConfig(selectedRule.id, config);
   }
 
@@ -53,20 +77,110 @@
 
   function updateMethod(method: string) {
     if (!selectedRule) return;
-    const config = { 
-      ...selectedRule.config, 
-      method: method.toUpperCase() as Rule['config']['method']
+    if (selectedRule.config.method === method) return;
+    const config = {
+      ...selectedRule.config,
+      method: method.toUpperCase() as Rule["config"]["method"],
     };
-    config.conditions = [{ key: '', value: '', response: '' }];
     ruleStore.updateRuleConfig(selectedRule.id, config);
   }
 
-  function updateMatchType(matchType: 'and' | 'or') {
+  function updateMatchType(matchType: "and" | "or") {
     if (!selectedRule) return;
+    if (selectedRule.config.matchType === matchType) return;
     const config = { ...selectedRule.config, matchType };
-    config.conditions = [{ key: '', value: '', response: '' }];
     ruleStore.updateRuleConfig(selectedRule.id, config);
   }
+
+  function handleResponseEditorSave() {
+    console.log("[info: 127]:", "保存编辑器并关闭");
+    if (editingConditionIndex >= 0 && selectedRule) {
+      const config = { ...selectedRule.config };
+      // @ts-ignore
+      config.conditions[editingConditionIndex].response =
+        mode === Mode.text
+          // @ts-ignore
+          ? responseContent.text
+          // @ts-ignore
+          : JSON.stringify(responseContent.json);
+      ruleStore.updateRuleConfig(selectedRule.id, config);
+      editingConditionIndex = -1;
+    }
+    showResponseEditor = false;
+  }
+
+  function handleNotify() {
+    selectedRule?.config.conditions.forEach((condition, index) => {
+      console.log("condition", condition);
+      if (condition.proxyMode === PROXY_MODE.NETWORK && condition.enabled) {
+        console.log(
+          "发出 sse 请求",
+          `${LOCAL_PREFIX}_${condition.ruleId}_${condition.key}_${condition.value}`,
+        );
+        notifyMessage({
+          storage_prefix: `${LOCAL_PREFIX}_${condition.ruleId}_${condition.key}_${condition.value}`,
+        }).then((res) => {
+          if (res) {
+            const newCondition = {...condition, response: res, proxyMode: PROXY_MODE.MOCK}
+            // 触发store更新以刷新UI
+            ruleStore.updateRuleConfigCondition({
+              ruleId: selectedRule!.id,
+              conditionIndex: index,
+              condition: newCondition,
+            });
+          } else {
+            condition.response = "";
+          }
+        });
+      }
+    });
+  }
+
+  function handleSwitchChange({
+    index,
+    status,
+    condition,
+  }: {
+    index: number;
+    status: boolean;
+    condition: RuleCondition;
+  }) {
+    const newCondition = { ...condition, enabled: status };
+    ruleStore.updateRuleConfigCondition({
+      ruleId: selectedRule!.id,
+      conditionIndex: index,
+      condition: newCondition,
+    });
+    console.log("Switch changed", status);
+  }
+
+  function handleSelectDropDown({
+    index,
+    selectOption,
+    condition,
+  }: {
+    index: number;
+    selectOption: { value: RuleProxyMode; label: string };
+    condition: RuleCondition;
+  }) {
+    const newCondition = { ...condition, proxyMode: selectOption.value };
+    ruleStore.updateRuleConfigCondition({
+      ruleId: selectedRule!.id,
+      conditionIndex: index,
+      condition: newCondition,
+    });
+  }
+
+  $effect(() => {
+    if (selectedRule && isFirst) {
+      handleNotify();
+      isFirst = false;
+    }
+  });
+
+  listenPageVisibility(handleNotify);
+
+  // 监听sse
 </script>
 
 <div class="rule-detail">
@@ -74,16 +188,14 @@
     <div class="detail-layout">
       <div class="detail-section basic-info">
         <div class="section-content">
-          <button class="add-btn" on:click={addCondition}>
-            添加条件
-          </button>
+          <button class="add-btn" onclick={addCondition}> 添加条件 </button>
           <div class="form-group">
             <label for="requestMethod">请求方式</label>
             <select
               id="requestMethod"
               class="form-input select-input"
               value={selectedRule.config.method.toLowerCase()}
-              on:change={(e) => updateMethod(e.currentTarget.value)}
+              onchange={(e) => updateMethod(e.currentTarget.value)}
             >
               <option value="get">GET</option>
               <option value="post">POST</option>
@@ -95,54 +207,97 @@
               id="matchType"
               class="form-input select-input"
               value={selectedRule.config.matchType}
-              on:change={(e) => updateMatchType(e.currentTarget.value as 'and' | 'or')}
+              onchange={(e) =>
+                updateMatchType(e.currentTarget.value as "and" | "or")}
             >
-              <option value="and">与（&）</option>
-              <option value="or">或（|）</option>
+            <option value="or">或</option>
+            <!-- <option value="and">与（&）</option> -->
             </select>
           </div>
           <div class="conditions-container">
             <fieldset>
               <legend>匹配条件</legend>
-              {#each selectedRule.config.conditions as condition, i}
-                <div class="condition-row" role="group" aria-labelledby="conditions-label">
-                  <input
-                    type="text"
-                    class="form-input condition-input"
-                    placeholder="Key"
-                    bind:value={condition.key}
-                    on:input={() => {
-                      if (selectedRule) {
-                        const config = { ...selectedRule.config };
-                        ruleStore.updateRuleConfig(selectedRule.id, config);
-                      }
-                    }}
+              {#each conditions as condition, i}
+                <div
+                  class="condition-row"
+                  role="group"
+                  aria-labelledby="conditions-label"
+                >
+                  <Switch
+                    size="small"
+                    checked={condition.enabled}
+                    onChange={(status: boolean) =>
+                      handleSwitchChange({ index: i, status, condition })}
                   />
-                  <input
-                    type="text"
-                    class="form-input condition-input"
-                    placeholder="Value"
-                    bind:value={condition.value}
-                    on:input={() => {
-                      if (selectedRule) {
-                        const config = { ...selectedRule.config };
-                        ruleStore.updateRuleConfig(selectedRule.id, config);
-                      }
-                    }}
-                  />
-                  <button
-                    class="edit-response-btn"
-                    on:click={() => openResponseEditor(i)}
-                  >
-                    编辑返回值
-                  </button>
-                  <button
-                    class="remove-btn"
-                    on:click={() => removeCondition(i)}
-                    disabled={selectedRule.config.conditions.length === 1}
-                  >
-                    删除
-                  </button>
+                  <div class="condition-inputs">
+                    <input
+                      type="text"
+                      class="form-input condition-input"
+                      placeholder="Key"
+                      bind:value={condition.key}
+                      oninput={() => {
+                        if (selectedRule) {
+                          const config = { ...selectedRule.config };
+                          ruleStore.updateRuleConfig(selectedRule.id, config);
+                        }
+                      }}
+                    />
+                    <input
+                      type="text"
+                      class="form-input condition-input"
+                      placeholder="Value"
+                      bind:value={condition.value}
+                      oninput={() => {
+                        if (selectedRule) {
+                          const config = { ...selectedRule.config };
+                          ruleStore.updateRuleConfig(selectedRule.id, config);
+                        }
+                      }}
+                    />
+                    <input
+                      type="text"
+                      class="form-input remark-input"
+                      placeholder="备注"
+                      bind:value={condition.remark}
+                      oninput={() => {
+                        if (selectedRule) {
+                          const config = { ...selectedRule.config };
+                          ruleStore.updateRuleConfig(selectedRule.id, config);
+                        }
+                      }}
+                    />
+                  </div>
+                  <div class="condition-actions">
+                    <Select
+                      options={[
+                        { value: "network", label: "网络模式" },
+                        { value: "mock", label: "mock模式" },
+                      ]}
+                      value={condition.proxyMode}
+                      class_="mode-select"
+                      onSelect={(selectOption) =>
+                        handleSelectDropDown({
+                          index: i,
+                          selectOption,
+                          condition,
+                        })}
+                      placeholder="选择模式"
+                    />
+                    <button
+                      class="edit-response-btn"
+                      onclick={() => openResponseEditor(i)}
+                    >
+                      {condition.proxyMode === PROXY_MODE.NETWORK
+                        ? "查看返回值"
+                        : "编辑返回值"}
+                    </button>
+                    <button
+                      class="remove-btn"
+                      onclick={() => removeCondition(i)}
+                    >
+                      删除
+                    </button>
+                  </div>
                 </div>
               {/each}
             </fieldset>
@@ -151,15 +306,15 @@
       </div>
 
       <ResponseEditor
-        bind:mode={mode}
+        bind:mode
         bind:showModal={showResponseEditor}
         bind:content={responseContent}
+        onClose={() => (showResponseEditor = false)}
+        onSave={handleResponseEditorSave}
       />
     </div>
   {:else}
-    <div class="no-selection">
-      选择一个规则查看详情
-    </div>
+    <div class="no-selection">选择一个规则查看详情</div>
   {/if}
 </div>
 
@@ -180,7 +335,6 @@
   .detail-section {
     background: #2a2a2a;
     border-radius: 8px;
-    overflow: hidden;
   }
 
   .section-content {
@@ -213,7 +367,6 @@
   }
 
   .form-input {
-    flex: 1;
     padding: 0.5rem;
     background-color: #2a2a2a;
     border: 1px solid #333;
@@ -260,12 +413,26 @@
     align-items: center;
   }
 
-  .condition-row .form-input {
+  .condition-inputs {
+    display: flex;
     flex: 1;
-    margin-right: 0.5rem;
+    gap: 0.5rem;
+  }
+
+  .condition-actions {
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+    margin-left: auto;
+  }
+  .remark-input,
+  .condition-input {
+    width: 200px;
+    border-color: gray;
   }
 
   .remove-btn {
+    flex-shrink: 0;
     padding: 0.5rem;
     background-color: #ff4d4f;
     border: none;
@@ -302,6 +469,7 @@
   }
 
   .edit-response-btn {
+    flex-shrink: 0;
     padding: 0.5rem;
     background-color: #1890ff;
     border: none;