+++ packages/whistle.interceptors/CHANGELOG.md
@@ -1,3 +1,27 @@
+# Changelog
+
+All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+
+### 0.0.4 (2025-05-23)
+
+
+### Features
+
+* 更新版本 ([f144aaf](https://github.com/lyx-jay/whistle.interceptors/commit/f144aaf11f021050b9a4353f799a710cc40f6863))
+* 支持多字段匹配 ([863cf1c](https://github.com/lyx-jay/whistle.interceptors/commit/863cf1cef5760a2d713e452b8add50a8e230ec0b))
+* 重构 ([f07881e](https://github.com/lyx-jay/whistle.interceptors/commit/f07881e5ac98b5fb337261e3e683ce686916de02))
+* add switch control network match condition ([84c65a8](https://github.com/lyx-jay/whistle.interceptors/commit/84c65a8528407d58bdc972ef7daeeea37953db21))
+* add toast ([ded7e13](https://github.com/lyx-jay/whistle.interceptors/commit/ded7e1341e07f200dc8d75e14b0669d3cc21a292))
+* hide and match mode ([72ea0ed](https://github.com/lyx-jay/whistle.interceptors/commit/72ea0edbe27a4e61bc63486fb6ec9537282c36f9))
+* perf ([ff0ba53](https://github.com/lyx-jay/whistle.interceptors/commit/ff0ba534c7efb52779878a014d166c34ca64adba))
+* response content auto record ([8796d52](https://github.com/lyx-jay/whistle.interceptors/commit/8796d52df2c74fe140a60523ccdcad5ddf27db93))
+* support ui ([1bf14d6](https://github.com/lyx-jay/whistle.interceptors/commit/1bf14d6b08a55f55e2425f84160c61d00871a0fe))
+
+
+### Bug Fixes
+
+* get list error ([c95903b](https://github.com/lyx-jay/whistle.interceptors/commit/c95903b90195a6aa1227531e21d736df3a9516b1))
+
 ## 0.0.3 (2025-05-05)
 
 
@@ -7,6 +31,3 @@
 * 支持多字段匹配 ([863cf1c](https://github.com/lyx-jay/whistle.interceptors/commit/863cf1cef5760a2d713e452b8add50a8e230ec0b))
 * 重构 ([f07881e](https://github.com/lyx-jay/whistle.interceptors/commit/f07881e5ac98b5fb337261e3e683ce686916de02))
 * support ui ([1bf14d6](https://github.com/lyx-jay/whistle.interceptors/commit/1bf14d6b08a55f55e2425f84160c61d00871a0fe))
-
-
-